requirements.txt
pytest==6.2.5
requests==2.25.1
pytest-xdist==2.1.0
allure-pytest==2.8.33
jsonschema==4.2.1
PyYAML==6.0
addict==2.4.0
mergedeep==1.3.4
pydantic==2.3.0
